Add keyboard support and fix TypeError bug
Allows keyboard to be used to enter numbers and decimal, choose an
operator, perform a computation, clear last entry and clear all.

Fix a bug where a String method is called on a Number. After compute()
is called, currentOperand stores the result which is a Number. When a
click event is fired while currentOperand is not empty after equals
button is clicked, appendNumber() is called and a String method is
invoked on currentOperand in the if statement.